requirements.txt
requests
responses
fastapi
jinja2
python-multipart
python-dotenv
uvicorn